move tied weights function to torch_utils
